Split node/service specific code
